Issue#887 merge spring cloud common with spring cloud module
