--- resilience4j-spring-cloud-common/build.gradle
@@ -1,13 +1,0 @@
-dependencies {
-    compile project(':resilience4j-spring')
-
-    compileOnly(libraries.spring_context)
-    compileOnly(libraries.spring_cloud_context2)
-    compileOnly(libraries.spring_boot2_autoconfigure)
-
-    testCompile(libraries.spring_context)
-    testCompile(libraries.spring_cloud_context2)
-}
-
-compileJava.dependsOn(processResources)
-ext.moduleName = 'io.github.resilience4j.spring-cloud-common'
