--- resilience4j-spring-cloud-common/src/main/java/io/github/resilience4j/circuitbreaker/autoconfigure/AbstractRefreshScopedCircuitBreakerConfiguration.java
@@ -1,47 +1,0 @@
-package io.github.resilience4j.circuitbreaker.autoconfigure;
-
-import io.github.resilience4j.circuitbreaker.CircuitBreaker;
-import io.github.resilience4j.circuitbreaker.CircuitBreakerRegistry;
-import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerConfiguration;
-import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerConfigurationProperties;
-import io.github.resilience4j.circuitbreaker.event.CircuitBreakerEvent;
-import io.github.resilience4j.common.CompositeCustomizer;
-import io.github.resilience4j.common.circuitbreaker.configuration.CircuitBreakerConfigCustomizer;
-import io.github.resilience4j.consumer.EventConsumerRegistry;
-import io.github.resilience4j.core.registry.RegistryEventConsumer;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.cloud.context.config.annotation.RefreshScope;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public abstract class AbstractRefreshScopedCircuitBreakerConfiguration {
-
-    protected final CircuitBreakerConfiguration circuitBreakerConfiguration;
-    protected final CircuitBreakerConfigurationProperties circuitBreakerProperties;
-
-    protected AbstractRefreshScopedCircuitBreakerConfiguration(
-        CircuitBreakerConfigurationProperties circuitBreakerProperties) {
-        this.circuitBreakerProperties = circuitBreakerProperties;
-        this.circuitBreakerConfiguration = new CircuitBreakerConfiguration(
-            circuitBreakerProperties);
-    }
-
-    /**
-     * @param eventConsumerRegistry the circuit breaker event consumer registry
-     * @return the RefreshScoped CircuitBreakerRegistry
-     */
-    @Bean
-    @RefreshScope
-    @ConditionalOnMissingBean
-    public CircuitBreakerRegistry circuitBreakerRegistry(
-        EventConsumerRegistry<CircuitBreakerEvent> eventConsumerRegistry,
-        RegistryEventConsumer<CircuitBreaker> circuitBreakerRegistryEventConsumer,
-        @Qualifier("compositeCircuitBreakerCustomizer") CompositeCustomizer<CircuitBreakerConfigCustomizer> compositeCircuitBreakerCustomizer) {
-        return circuitBreakerConfiguration
-            .circuitBreakerRegistry(eventConsumerRegistry, circuitBreakerRegistryEventConsumer,
-                compositeCircuitBreakerCustomizer);
-    }
-
-}
