+++ resilience4j-spring-cloud2/src/main/java/io/github/resilience4j/retry/autoconfigure/RefreshScopedRetryAutoConfiguration.java
@@ -1,17 +1,51 @@
 package io.github.resilience4j.retry.autoconfigure;
 
+import io.github.resilience4j.common.CompositeCustomizer;
+import io.github.resilience4j.common.retry.configuration.RetryConfigCustomizer;
+import io.github.resilience4j.consumer.EventConsumerRegistry;
+import io.github.resilience4j.core.registry.RegistryEventConsumer;
 import io.github.resilience4j.retry.Retry;
+import io.github.resilience4j.retry.RetryRegistry;
+import io.github.resilience4j.retry.configure.RetryConfiguration;
+import io.github.resilience4j.retry.configure.RetryConfigurationProperties;
+import io.github.resilience4j.retry.event.RetryEvent;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.cloud.autoconfigure.RefreshAutoConfiguration;
 import org.springframework.cloud.context.scope.refresh.RefreshScope;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ConditionalOnClass({Retry.class, RefreshScope.class})
 @AutoConfigureAfter(RefreshAutoConfiguration.class)
 @AutoConfigureBefore(RetryAutoConfiguration.class)
-public class RefreshScopedRetryAutoConfiguration extends AbstractRefreshScopedRetryConfiguration {
+public class RefreshScopedRetryAutoConfiguration {
+
+    private final RetryConfiguration retryConfiguration;
+
+    protected RefreshScopedRetryAutoConfiguration() {
+        this.retryConfiguration = new RetryConfiguration();
+    }
+
+    /**
+     * @param retryConfigurationProperties retry spring configuration properties
+     * @param retryEventConsumerRegistry   the retry event consumer registry
+     * @return the RefreshScoped RetryRegistry
+     */
+    @Bean
+    @org.springframework.cloud.context.config.annotation.RefreshScope
+    @ConditionalOnMissingBean
+    public RetryRegistry retryRegistry(RetryConfigurationProperties retryConfigurationProperties,
+                                       EventConsumerRegistry<RetryEvent> retryEventConsumerRegistry,
+                                       RegistryEventConsumer<Retry> retryRegistryEventConsumer,
+                                       @Qualifier("compositeRetryCustomizer") CompositeCustomizer<RetryConfigCustomizer> compositeRetryCustomizer) {
+        return retryConfiguration
+            .retryRegistry(retryConfigurationProperties, retryEventConsumerRegistry,
+                retryRegistryEventConsumer, compositeRetryCustomizer);
+    }
 
 }
