+++ resilience4j-spring-cloud2/src/test/java/io/github/resilience4j/circuitbreaker/autoconfigure/RefreshScopedCircuitBreakerConfigurationTest.java
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package io.github.resilience4j.ratelimiter.autoconfigure;
+package io.github.resilience4j.circuitbreaker.autoconfigure;
 
+import io.github.resilience4j.circuitbreaker.configure.CircuitBreakerConfigurationProperties;
 import io.github.resilience4j.common.CompositeCustomizer;
 import io.github.resilience4j.consumer.DefaultEventConsumerRegistry;
 import io.github.resilience4j.core.registry.CompositeRegistryEventConsumer;
-import io.github.resilience4j.ratelimiter.configure.RateLimiterConfigurationProperties;
-import io.github.resilience4j.retry.autoconfigure.AbstractRefreshScopedRetryConfiguration;
 import org.junit.Test;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Bean;
@@ -31,26 +30,31 @@ import java.util.Collections;
 import static java.util.Collections.emptyList;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class AbstractRefreshScopedRateLimiterConfigurationTest {
+public class RefreshScopedCircuitBreakerConfigurationTest {
 
     @Test
-    public void testRefreshScopedRetryConfig() {
-        Arrays.stream(AbstractRefreshScopedRetryConfiguration.class.getMethods())
+    public void testRefreshScopedCircuitBreakerConfig() {
+        Arrays.stream(RefreshScopedCircuitBreakerAutoConfiguration.class.getMethods())
             .filter(method -> method.isAnnotationPresent(Bean.class))
             .forEach(method -> assertThat(method.isAnnotationPresent(RefreshScope.class)).isTrue());
     }
 
     @Test
-    public void testRateLimiterCloudCommonConfig() {
-        RateLimiterConfig rateLimiterConfig = new RateLimiterConfig();
+    public void testCircuitBreakerCloudCommonConfig() {
+        CircuitBreakerConfig circuitBreakerConfig = new CircuitBreakerConfig(
+            new CircuitBreakerConfigurationProperties());
 
-        assertThat(rateLimiterConfig.rateLimiterRegistry(
-            new RateLimiterConfigurationProperties(), new DefaultEventConsumerRegistry<>(),
-            new CompositeRegistryEventConsumer<>(emptyList()), new CompositeCustomizer<>(
-                Collections.emptyList()))).isNotNull();
+        assertThat(circuitBreakerConfig.circuitBreakerRegistry(
+            new DefaultEventConsumerRegistry<>(),
+            new CompositeRegistryEventConsumer<>(emptyList()),
+            new CompositeCustomizer<>(Collections.emptyList())))
+            .isNotNull();
     }
 
-    static class RateLimiterConfig extends AbstractRefreshScopedRateLimiterConfiguration {
+    static class CircuitBreakerConfig extends RefreshScopedCircuitBreakerAutoConfiguration {
 
+        CircuitBreakerConfig(CircuitBreakerConfigurationProperties circuitBreakerProperties) {
+            super(circuitBreakerProperties);
+        }
     }
 }
