+++ settings.gradle
@@ -16,7 +16,6 @@ include 'resilience4j-annotations'
 include 'resilience4j-spring'
 include 'resilience4j-spring-boot2'
 include 'resilience4j-spring-cloud2'
-include 'resilience4j-spring-cloud-common'
 include 'resilience4j-retrofit'
 include 'resilience4j-feign'
 include 'resilience4j-ratpack'
